Refactor namespace of for the data plugin
