+++ braid-data-plugin/src/main/java/mx/sugus/braid/plugins/data/producers/DirectedInterface.java
@@ -1,33 +1,37 @@
-package mx.sugus.braid.plugins.data;
+package mx.sugus.braid.plugins.data.producers;
 
 import java.util.Collections;
 import java.util.List;
 import mx.sugus.braid.core.plugin.ShapeCodegenState;
-import mx.sugus.braid.jsyntax.ClassName;
-import mx.sugus.braid.jsyntax.ConstructorMethodSyntax;
-import mx.sugus.braid.jsyntax.EnumSyntax;
+import mx.sugus.braid.jsyntax.AbstractMethodSyntax;
 import mx.sugus.braid.jsyntax.FieldSyntax;
+import mx.sugus.braid.jsyntax.InterfaceSyntax;
 import mx.sugus.braid.jsyntax.MethodSyntax;
+import mx.sugus.braid.jsyntax.TypeName;
 import mx.sugus.braid.jsyntax.TypeSyntax;
+import software.amazon.smithy.model.shapes.MemberShape;
 
-public interface DirectedEnum extends DirectiveToTypeSyntax {
-    default ClassName className(ShapeCodegenState state) {
-        return ClassName.toClassName(Utils.toJavaTypeName(state, state.shape()));
-    }
+public interface DirectedInterface extends DirectiveToTypeSyntax {
+    TypeName className(ShapeCodegenState state);
 
-    EnumSyntax.Builder typeSpec(ShapeCodegenState state);
+    InterfaceSyntax.Builder typeSpec(ShapeCodegenState state);
 
     default List<FieldSyntax> extraFields(ShapeCodegenState state) {
         return Collections.emptyList();
     }
 
-    default List<ConstructorMethodSyntax> constructors(ShapeCodegenState state) {
+    default List<AbstractMethodSyntax> abstractMethodsFor(ShapeCodegenState state, MemberShape member) {
+        return Collections.emptyList();
+    }
+
+    default List<AbstractMethodSyntax> extraAbstractMethods(ShapeCodegenState state) {
         return Collections.emptyList();
     }
 
     default List<MethodSyntax> extraMethods(ShapeCodegenState state) {
         return Collections.emptyList();
     }
+
 
     default List<DirectiveToTypeSyntax> innerTypes(ShapeCodegenState state) {
         return Collections.emptyList();
@@ -39,8 +43,12 @@ public interface DirectedEnum extends DirectiveToTypeSyntax {
         for (var field : extraFields(state)) {
             builder.addField(field);
         }
-
-        for (var method : constructors(state)) {
+        for (var member : state.shape().members()) {
+            for (var method : abstractMethodsFor(state, member)) {
+                builder.addMethod(method);
+            }
+        }
+        for (var method : extraAbstractMethods(state)) {
             builder.addMethod(method);
         }
         for (var method : extraMethods(state)) {
