+++ braid-data-plugin/src/main/java/mx/sugus/braid/plugins/data/producers/Utils.java
@@ -1,4 +1,4 @@
-package mx.sugus.braid.plugins.data;
+package mx.sugus.braid.plugins.data.producers;
 
 import java.util.ArrayList;
 import java.util.List;
