+++ braid-serde-node-plugin/src/main/java/mx/sugus/braid/plugins/serde/node/ClassAddToNodeTransformer.java
@@ -12,8 +12,8 @@ import mx.sugus.braid.jsyntax.CodeBlock;
 import mx.sugus.braid.jsyntax.MethodSyntax;
 import mx.sugus.braid.jsyntax.block.BodyBuilder;
 import mx.sugus.braid.jsyntax.ext.JavadocExt;
-import mx.sugus.braid.plugins.data.StructureJavaProducer;
-import mx.sugus.braid.plugins.data.Utils;
+import mx.sugus.braid.plugins.data.producers.StructureJavaProducer;
+import mx.sugus.braid.plugins.data.producers.Utils;
 import mx.sugus.braid.traits.ConstTrait;
 import mx.sugus.braid.traits.JavaTrait;
 import software.amazon.smithy.model.node.ArrayNode;
