+++ braid-serde-node-plugin/src/main/java/mx/sugus/braid/plugins/serde/node/InterfaceAddFromNodeTransformer.java
@@ -12,8 +12,8 @@ import mx.sugus.braid.jsyntax.DefaultCaseClause;
 import mx.sugus.braid.jsyntax.InterfaceSyntax;
 import mx.sugus.braid.jsyntax.MethodSyntax;
 import mx.sugus.braid.jsyntax.SwitchStatement;
-import mx.sugus.braid.plugins.data.StructureInterfaceJavaProducer;
-import mx.sugus.braid.plugins.data.Utils;
+import mx.sugus.braid.plugins.data.producers.StructureInterfaceJavaProducer;
+import mx.sugus.braid.plugins.data.producers.Utils;
 import mx.sugus.braid.traits.ConstTrait;
 import software.amazon.smithy.model.node.Node;
 import software.amazon.smithy.model.node.ObjectNode;
