+++ braid-serde-node-plugin/src/main/java/mx/sugus/braid/plugins/serde/node/InterfaceAddToNodeTransformer.java
@@ -5,7 +5,7 @@ import mx.sugus.braid.core.plugin.ShapeCodegenState;
 import mx.sugus.braid.core.plugin.ShapeTaskTransformer;
 import mx.sugus.braid.core.plugin.TypeSyntaxResult;
 import mx.sugus.braid.jsyntax.InterfaceSyntax;
-import mx.sugus.braid.plugins.data.StructureInterfaceJavaProducer;
+import mx.sugus.braid.plugins.data.producers.StructureInterfaceJavaProducer;
 import software.amazon.smithy.model.node.ToNode;
 
 public final class InterfaceAddToNodeTransformer implements ShapeTaskTransformer<TypeSyntaxResult> {
