+++ braid-syntax-plugin/src/main/java/mx/sugus/braid/plugins/syntax/InterfaceSyntaxAddAcceptVisitorTransformer.java
@@ -9,8 +9,8 @@ import mx.sugus.braid.jsyntax.ClassName;
 import mx.sugus.braid.jsyntax.InterfaceSyntax;
 import mx.sugus.braid.jsyntax.ParameterizedTypeName;
 import mx.sugus.braid.jsyntax.TypeVariableTypeName;
-import mx.sugus.braid.plugins.data.StructureInterfaceJavaProducer;
-import mx.sugus.braid.plugins.data.Utils;
+import mx.sugus.braid.plugins.data.producers.StructureInterfaceJavaProducer;
+import mx.sugus.braid.plugins.data.producers.Utils;
 import mx.sugus.braid.traits.InterfaceTrait;
 import software.amazon.smithy.model.shapes.ShapeId;
