+++ braid-syntax-plugin/src/main/java/mx/sugus/braid/plugins/syntax/SyntaxVisitorJavaProducer.java
@@ -14,7 +14,7 @@ import mx.sugus.braid.jsyntax.MethodSyntax;
 import mx.sugus.braid.jsyntax.ParameterizedTypeName;
 import mx.sugus.braid.jsyntax.TypeSyntax;
 import mx.sugus.braid.jsyntax.TypeVariableTypeName;
-import mx.sugus.braid.plugins.data.Utils;
+import mx.sugus.braid.plugins.data.producers.Utils;
 import mx.sugus.braid.traits.InterfaceTrait;
 import software.amazon.smithy.model.Model;
 import software.amazon.smithy.model.shapes.Shape;
